+++ src/lib/rsc-utils.ts
@@ -7,7 +7,7 @@ globalThis.__webpack_chunk_load__ = (id) => import(id).then((m) => globalThis.__
 globalThis.__webpack_require__ = (id) => globalThis.__waku_module_cache__.get(id);`;
 
 export const generatePrefetchCode = (
-  basePath: string,
+  basePrefix: string,
   entryItemsIterable: Iterable<readonly [rscId: string, props: unknown]>,
   moduleIds: Iterable<string>
 ) => {
@@ -29,7 +29,7 @@ ${rscIds
           const searchParams = new URLSearchParams();
           searchParams.set("props", serializedProps);
           return [
-            `'${serializedProps}': fetch('${basePath}${rscId}/${searchParams}')`,
+            `'${serializedProps}': fetch('${basePrefix}${rscId}/${searchParams}')`,
           ];
         })
         .join(",") +