+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Change Log
 
 ## [Unreleased]
+### Changed
+- fix: rscPrefix config without trailing slash #50
 
 ## [0.11.0] - 2023-05-25
 ### Changed